% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapCC.R
\name{BootstrapCC}
\alias{BootstrapCC}
\title{BootstrapCC}
\usage{
BootstrapCC(
  d = NULL,
  maxK = 10,
  reps = 100,
  clusterAlg = "hc",
  title = "untitled_consensus_cluster",
  innerLinkage = "average",
  finalLinkage = "average",
  distance = "pearson",
  ml = NULL,
  tmyPal = NULL,
  seed = NULL,
  plot = NULL,
  writeTable = FALSE,
  verbose = F,
  corUse = "everything",
  nCores = 1L
)
}
\arguments{
\item{d}{data to be clustered; either a data matrix where columns=items/samples and rows are features.  For example, a gene expression matrix of genes in rows and microarrays in columns, or ExpressionSet object, or a distance object (only for cases of no feature resampling)}

\item{maxK}{(numeric) maximum number of clusters to test (default 10)}

\item{reps}{(numeric) maximum number of subsamplings iterations (default 100)}

\item{clusterAlg}{character value. cluster algorithm.  'hc' heirarchical (hclust), 'pam' for paritioning around medoids, 'km' for k-means upon data matrix, 'kmdist' for k-means upon distance matrices (former km option), or a function that returns a clustering.  See example and vignette for more details.}

\item{title}{character value for output directory. Directory is created only if plot is not NULL or writeTable is TRUE. This title can be an abosulte or relative path.}

\item{innerLinkage}{heirarchical linkage method for subsampling.}

\item{finalLinkage}{eirarchical linkage method for consensus matrix.}

\item{distance}{character value. 'pearson': (1 - Pearson correlation), 'spearman' (1 - Spearman correlation), 'euclidean', 'binary', 'maximum', 'canberra', 'minkowski" or custom distance function.}

\item{ml}{optional. prior result, if supplied then only do graphics and tables.}

\item{tmyPal}{optional character vector of colors for consensus matrix}

\item{seed}{optional numerical value.  sets random seed for reproducible results.}

\item{plot}{character value. NULL - print to screen, 'pdf', 'png', 'pngBMP' for bitmap png, helpful for large datasets.}

\item{writeTable}{logical value. TRUE - write ouput and log to csv.}

\item{corUse}{optional character value.  specifies how to handle missing data in correlation distances 'everything','pairwise.complete.obs', 'complete.obs' see cor() for description.}

\item{nCores}{(numeric) number of cpu cores to use (default 1)}
}
\value{
ConsensusClusterPlus returns a list of length maxK.  Each element is a list containing consensusMatrix (numerical matrix), consensusTree (hclust), consensusClass (consensus class asssignments).  ConsensusClusterPlus also produces images.
calcICL returns a list of two elements clusterConsensus and itemConsensus corresponding to cluster-consensus and item-consensus.  See Monti, et al (2003) for formulas.
}
\description{
A fast Bootstrap implementation of consensus clustering plus algorithm
}
\examples{
data(iris)
BootstrapCC(t(data.matrix(iris[,1:4])))

}
\references{
Please cite the ConsensusClusterPlus publication, below, if you use ConsensusClusterPlus in a publication or presentation:

 Wilkerson, M.D., Hayes, D.N. (2010). ConsensusClusterPlus: a class discovery tool with confidence assessments and item tracking. Bioinformatics, 2010 Jun 15;26(12):1572-3.

 Original description of the Consensus Clustering method:

 Monti, S., Tamayo, P., Mesirov, J., Golub, T. (2003) Consensus Clustering:
 A Resampling-Based Method for Class Discovery and Visualization of Gene
 Expression Microarray Data. Machine Learning, 52, 91-118.
}
\seealso{
[ConsensusClusterPlus::ConsensusClusterPlus()] which this function is build on top.
}
\author{
Fernández Elmer efernandez@cidie.ucc.edu.ar
}
